Add initial Slack plugin
The Slack plugin uses Markdown but must render a complex BlockKit
structure to the Slack API. To make this work, allow output drivers to
also get a list of views with their rendered outputs and blobs tied to
them. Also allow blobs to have titles. Still some awkardness here to
sort out, but the basics are in place.