Adding dev container setup
This is starting to get a bit more complicated. Trying to simplify some
of the dependencies. The main issue is how to account for the fact that
plugins require dependencies that have dependencies on binaries like
mysql-client. And docs need the plugins to render autodocs. It's
possible to not require the dependencies but this is hard to do in a
generic manner.