Add wait mechanism for run/exec race condition
It's very possible for a container to start but be stuck doing things in
its entrypoint while a request for an exec comes in. Not sure how Docker
could even solve this, but this adds a mechanism to signal that the
kpireport container is really "ready" - the issue is that we
lazy-install all the local packages when the container starts, because
they have to be locally mounted.