Handle non-numeric extra columns more gracefully
When passing in a dataframe with multiple non-numeric columns, most
plot functions would fail because the types are all wrong. In general
this is probably a mistake we should try to help prevent. Log an error
and silently drop the data. There is some trickery involved in making
sure by dropping the data we don't mess things up too much.