[static] PNG output support
This updates the static plugin to support outputting a PNG file
instead of an HTML report. It uses wkhtmltoimage in the background,
via the ``imgkit`` Python package.

The output PNG is cropped such that there is no border/background
on either side; the PNG takes up the full report width.