Use tz-aware datetimes for start/end dates
Prior, the report would use whatever the default return of
datetime.now() was, which was often incorrect, and at the very least
was not configurable. This updates it to properly set timezones for
all times (derived from a local timezone). It also adds a ``timezone``
argument, which can be used to override the system timezone for some
reason (useful when running in a container without TZ set.)