Add Twitter engagement view
This adds a new Twitter view, which plots "engagement" on a Twitter
account. It looks up the list of tweets for the target user and plots
the tweets' like/reply/retweet counts. The highest-liked tweet is called
out with a bit of text for additional context.

This plugin extends the Plot plugin and sets up some defaults, and uses
the ``post_plot`` hook to add the tweet callouts.