Look up plugin templates in entire class hierarchy
The old plugin lookup function would only look at the implementing
class's module for templates. However, when subclassing, we want
to use the subclass' templates when found. The plugin loading is
now aware of class hierarchies and will construct a lookup chain
starting from the concrete class and going down to the base View.